change strcture: train test functions in resnet class
